Cinemas Controller Updated

+++ Controllers/CinemasController.cs
@@ -1,46 +1,36 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
-using Microsoft.AspNetCore.Mvc;
-using Microsoft.AspNetCore.Mvc.Rendering;
-using Microsoft.EntityFrameworkCore;
+﻿using HotelManagement.Areas.Identity.Data;
 using HotelManagement.Data;
 using HotelManagement.Models;
+using Microsoft.AspNetCore.Authorization;
+using Microsoft.AspNetCore.Identity;
+using Microsoft.AspNetCore.Mvc;
+using Microsoft.EntityFrameworkCore;
+using System.Linq;
+using System.Security.Claims;
+using System.Threading.Tasks;
 
 namespace HotelManagement.Controllers
 {
+    [Authorize] // Ensure that the user is authenticated
     public class CinemasController : Controller
     {
         private readonly HotelContext _context;
+        private readonly UserManager<HotelUser> _userManager;
 
-        public CinemasController(HotelContext context)
+        public CinemasController(HotelContext context, UserManager<HotelUser> userManager)
         {
             _context = context;
+            _userManager = userManager;
         }
 
         // GET: Cinema
         public async Task<IActionResult> Index()
         {
-            return View(await _context.Cinema.ToListAsync());
-        }
-
-        // GET: Cinema/Details/5
-        public async Task<IActionResult> Details(int? id)
-        {
-            if (id == null || _context.Cinema == null)
-            {
-                return NotFound();
-            }
-
-            var cinema = await _context.Cinema
-                .FirstOrDefaultAsync(m => m.CinemaID == id);
-            if (cinema == null)
-            {
-                return NotFound();
-            }
-
-            return View(cinema);
+            var userId = User.FindFirstValue(ClaimTypes.NameIdentifier); // Get the current user's ID
+            var cinemaBookings = await _context.Cinema
+                                               .Where(c => c.UserId == userId) // Filter by user
+                                               .ToListAsync();
+            return View(cinemaBookings);
         }
 
         // GET: Cinema/Create
@@ -50,14 +40,13 @@ namespace HotelManagement.Controllers
         }
 
         // POST: Cinema/Create
-        // To protect from overposting attacks, enable the specific properties you want to bind to.
-        // For more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Create([Bind("CinemaID,GuestName,BookingTime,MovieTitle,NumberOfTickets,Confirmed")] Cinema cinema)
+        public async Task<IActionResult> Create([Bind("GuestName,BookingTime,MovieTitle,NumberOfTickets")] Cinema cinema)
         {
             if (ModelState.IsValid)
             {
+                cinema.UserId = User.FindFirstValue(ClaimTypes.NameIdentifier); // Assign current user's ID
                 _context.Add(cinema);
                 await _context.SaveChangesAsync();
                 return RedirectToAction(nameof(Index));
@@ -68,35 +57,29 @@ namespace HotelManagement.Controllers
         // GET: Cinema/Edit/5
         public async Task<IActionResult> Edit(int? id)
         {
-            if (id == null || _context.Cinema == null)
-            {
-                return NotFound();
-            }
+            if (id == null) return NotFound();
 
             var cinema = await _context.Cinema.FindAsync(id);
-            if (cinema == null)
+            if (cinema == null || cinema.UserId != User.FindFirstValue(ClaimTypes.NameIdentifier))
             {
-                return NotFound();
+                return NotFound(); // Prevent access if the user does not own the booking
             }
+
             return View(cinema);
         }
 
         // POST: Cinema/Edit/5
-        // To protect from overposting attacks, enable the specific properties you want to bind to.
-        // For more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> Edit(int id, [Bind("CinemaID,GuestName,BookingTime,MovieTitle,NumberOfTickets,Confirmed")] Cinema cinema)
         {
-            if (id != cinema.CinemaID)
-            {
-                return NotFound();
-            }
+            if (id != cinema.CinemaID) return NotFound();
 
             if (ModelState.IsValid)
             {
                 try
                 {
+                    cinema.UserId = User.FindFirstValue(ClaimTypes.NameIdentifier); // Ensure the correct user is editing
                     _context.Update(cinema);
                     await _context.SaveChangesAsync();
                 }
@@ -106,10 +89,7 @@ namespace HotelManagement.Controllers
                     {
                         return NotFound();
                     }
-                    else
-                    {
-                        throw;
-                    }
+                    throw;
                 }
                 return RedirectToAction(nameof(Index));
             }
@@ -119,16 +99,13 @@ namespace HotelManagement.Controllers
         // GET: Cinema/Delete/5
         public async Task<IActionResult> Delete(int? id)
         {
-            if (id == null || _context.Cinema == null)
-            {
-                return NotFound();
-            }
+            if (id == null) return NotFound();
 
             var cinema = await _context.Cinema
                 .FirstOrDefaultAsync(m => m.CinemaID == id);
-            if (cinema == null)
+            if (cinema == null || cinema.UserId != User.FindFirstValue(ClaimTypes.NameIdentifier))
             {
-                return NotFound();
+                return NotFound(); // Prevent access if the user does not own the booking
             }
 
             return View(cinema);
@@ -139,18 +116,13 @@ namespace HotelManagement.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
-            if (_context.Cinema == null)
-            {
-                return Problem("Entity set 'HotelContext.Cinema'  is null.");
-            }
-
             var cinema = await _context.Cinema.FindAsync(id);
-            if (cinema != null)
+            if (cinema != null && cinema.UserId == User.FindFirstValue(ClaimTypes.NameIdentifier))
             {
                 _context.Cinema.Remove(cinema);
+                await _context.SaveChangesAsync();
             }
 
-            await _context.SaveChangesAsync();
             return RedirectToAction(nameof(Index));
         }
